extract closest holiday component
